+++ test/Engine-test.js
@@ -44,6 +44,7 @@ describe('Engine', () => {
 
     it('throws if unsupported source is passed', (done) => {
       const engine = Bpmn.Engine({
+        // @ts-expect-error
         source: {},
       });
 
@@ -66,6 +67,7 @@ describe('Engine', () => {
 
     it('but not function', (done) => {
       const engine = Bpmn.Engine({
+        // @ts-expect-error
         source() {},
       });
 
@@ -123,6 +125,7 @@ describe('Engine', () => {
       expect(() =>
         Bpmn.Engine({
           source: factory.valid(),
+          // @ts-expect-error
           Logger,
         })
       ).to.throw(TypeError, 'Logger is not a function');
@@ -306,7 +309,7 @@ describe('Engine', () => {
           input: 0,
         },
         services: {
-          get(context, next) {
+          get(_context, next) {
             next(new Error('Inner error'));
           },
         },
@@ -433,7 +436,7 @@ describe('Engine', () => {
         name: 'end test',
         source,
         services: {
-          get: (context, next) => {
+          get: (_context, next) => {
             next(new Error('Inner error'));
           },
         },
@@ -461,7 +464,7 @@ describe('Engine', () => {
         name: 'end test',
         source,
         services: {
-          get: (context, next) => {
+          get: (_context, next) => {
             next(new Error('Inner error'));
           },
         },
@@ -482,6 +485,7 @@ describe('Engine', () => {
 
       try {
         await engine.execute({
+          // @ts-expect-error
           listener: {},
         });
       } catch (e) {
@@ -1156,6 +1160,7 @@ describe('Engine', () => {
 
     it('execution api returns activities in a postponed state', async () => {
       const listener = new EventEmitter();
+      /** @type {import('bpmn-engine').Execution} */
       let engineApi;
       listener.once('wait', (_, api) => {
         engineApi = api;
@@ -1230,6 +1235,7 @@ describe('Engine', () => {
         sourceContext: updateContext,
       });
 
+      /** @type {import('bpmn-engine').Execution} */
       let engineApi;
       listener.once('wait', (_, api) => {
         engineApi = api;